Don't fail update upon error fetching the incumbent worker's import scripts.
To include imported scripts in the byte-for-byte comparison, the Update
algorithm fetches the incumbent worker's scripts in order to
detect if there is a difference. But it would abort the update upon
MIME type error of if fetching the script failed. This is undesirable
because the new worker might not use these scripts at all, e.g.,
the scripts may have legitimitely been removed from the server.

To fix this, just set the updated flag upon detecting a difference
and continue with the Update.

Addresses issue #1374.